btstack_event: regenerate

+++ src/btstack_event.h
@@ -5818,7 +5818,7 @@ static inline uint8_t gap_subevent_big_created_get_num_bis(const uint8_t * event
  * @note: btstack_type C
  */
 static inline uint16_t gap_subevent_big_created_get_bis_con_handles(const uint8_t * event, uint8_t index){
-    return little_endian_read_16(event, 6 + (2 * index));
+    return little_endian_read_16(event, 6 + (2 * (int) index));
 }
 
 /**
@@ -5866,7 +5866,7 @@ static inline uint8_t gap_subevent_big_sync_created_get_num_bis(const uint8_t * 
  * @note: btstack_type C
  */
 static inline uint16_t gap_subevent_big_sync_created_get_bis_con_handles(const uint8_t * event, uint8_t index){
-    return little_endian_read_16(event, 6 + (2 * index));
+    return little_endian_read_16(event, 6 + (2 * (int) index));
 }
 
 /**
@@ -5914,7 +5914,7 @@ static inline uint8_t gap_subevent_cig_created_get_num_cis(const uint8_t * event
  * @note: btstack_type C
  */
 static inline uint16_t gap_subevent_cig_created_get_cis_con_handles(const uint8_t * event, uint8_t index){
-    return little_endian_read_16(event, 6 + (2 * index));
+    return little_endian_read_16(event, 6 + (2 * (int) index));
 }
 
 /**
@@ -14340,7 +14340,7 @@ static inline const uint8_t * gattservice_subevent_ascs_server_metadata_get_prog
  * @note: btstack_type 3
  */
 static inline uint32_t gattservice_subevent_ascs_server_metadata_get_language_code(const uint8_t * event){
-    return little_endian_read_24(event, 12u + event[11]);
+    return little_endian_read_24(event, (int)(12u + event[11]));
 }
 /**
  * @brief Get field ccids_num from event GATTSERVICE_SUBEVENT_ASCS_SERVER_METADATA
@@ -14394,7 +14394,7 @@ static inline const uint8_t * gattservice_subevent_ascs_server_metadata_get_prog
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_ascs_server_metadata_get_extended_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u]);
+    return little_endian_read_16(event, (int)(12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u]));
 }
 /**
  * @brief Get field extended_metadata_value_length from event GATTSERVICE_SUBEVENT_ASCS_SERVER_METADATA
@@ -14421,7 +14421,7 @@ static inline const uint8_t * gattservice_subevent_ascs_server_metadata_get_exte
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_ascs_server_metadata_get_vendor_specific_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u]);
+    return little_endian_read_16(event, (int)(12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u]));
 }
 /**
  * @brief Get field vendor_specific_metadata_value_length from event GATTSERVICE_SUBEVENT_ASCS_SERVER_METADATA
@@ -14784,7 +14784,7 @@ static inline const uint8_t * gattservice_subevent_ascs_client_metadata_get_prog
  * @note: btstack_type 3
  */
 static inline uint32_t gattservice_subevent_ascs_client_metadata_get_language_code(const uint8_t * event){
-    return little_endian_read_24(event, 12u + event[11]);
+    return little_endian_read_24(event, (int)(12u + event[11]));
 }
 /**
  * @brief Get field ccids_num from event GATTSERVICE_SUBEVENT_ASCS_CLIENT_METADATA
@@ -14838,7 +14838,7 @@ static inline const uint8_t * gattservice_subevent_ascs_client_metadata_get_prog
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_ascs_client_metadata_get_extended_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u]);
+    return little_endian_read_16(event, (int)(12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u]));
 }
 /**
  * @brief Get field extended_metadata_value_length from event GATTSERVICE_SUBEVENT_ASCS_CLIENT_METADATA
@@ -14865,7 +14865,7 @@ static inline const uint8_t * gattservice_subevent_ascs_client_metadata_get_exte
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_ascs_client_metadata_get_vendor_specific_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u]);
+    return little_endian_read_16(event, (int)(12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u + 1u + event[12u + event[11] + 3u + 1u + event[12u + event[11] + 3u] + 1u] + 2u]));
 }
 /**
  * @brief Get field vendor_specific_metadata_value_length from event GATTSERVICE_SUBEVENT_ASCS_CLIENT_METADATA
@@ -15422,7 +15422,7 @@ static inline const uint8_t * gattservice_subevent_bass_client_notify_receive_st
  * @note: btstack_type 3
  */
 static inline uint32_t gattservice_subevent_bass_client_notify_receive_state_subgroup_get_language_code(const uint8_t * event){
-    return little_endian_read_24(event, 16u + event[15]);
+    return little_endian_read_24(event, (int)(16u + event[15]));
 }
 /**
  * @brief Get field ccids_num from event GATTSERVICE_SUBEVENT_BASS_CLIENT_NOTIFY_RECEIVE_STATE_SUBGROUP
@@ -15476,7 +15476,7 @@ static inline const uint8_t * gattservice_subevent_bass_client_notify_receive_st
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_bass_client_notify_receive_state_subgroup_get_extended_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u]);
+    return little_endian_read_16(event, (int)(16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u]));
 }
 /**
  * @brief Get field extended_metadata_value_length from event GATTSERVICE_SUBEVENT_BASS_CLIENT_NOTIFY_RECEIVE_STATE_SUBGROUP
@@ -15503,7 +15503,7 @@ static inline const uint8_t * gattservice_subevent_bass_client_notify_receive_st
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_bass_client_notify_receive_state_subgroup_get_vendor_specific_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u] + 2u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u] + 2u]);
+    return little_endian_read_16(event, (int)(16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u] + 2u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u + 1u + event[16u + event[15] + 3u + 1u + event[16u + event[15] + 3u] + 1u] + 2u]));
 }
 /**
  * @brief Get field vendor_specific_metadata_value_length from event GATTSERVICE_SUBEVENT_BASS_CLIENT_NOTIFY_RECEIVE_STATE_SUBGROUP
@@ -15898,7 +15898,7 @@ static inline const uint8_t * gattservice_subevent_pacs_client_pack_record_get_p
  * @note: btstack_type 3
  */
 static inline uint32_t gattservice_subevent_pacs_client_pack_record_get_language_code(const uint8_t * event){
-    return little_endian_read_24(event, 28u + event[27]);
+    return little_endian_read_24(event, (int)(28u + event[27]));
 }
 /**
  * @brief Get field ccids_num from event GATTSERVICE_SUBEVENT_PACS_CLIENT_PACK_RECORD
@@ -15952,7 +15952,7 @@ static inline const uint8_t * gattservice_subevent_pacs_client_pack_record_get_p
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_pacs_client_pack_record_get_extended_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u]);
+    return little_endian_read_16(event, (int)(28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u]));
 }
 /**
  * @brief Get field extended_metadata_value_length from event GATTSERVICE_SUBEVENT_PACS_CLIENT_PACK_RECORD
@@ -15979,7 +15979,7 @@ static inline const uint8_t * gattservice_subevent_pacs_client_pack_record_get_e
  * @note: btstack_type 2
  */
 static inline uint16_t gattservice_subevent_pacs_client_pack_record_get_vendor_specific_metadata_type(const uint8_t * event){
-    return little_endian_read_16(event, 28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u] + 2u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u] + 2u]);
+    return little_endian_read_16(event, (int)(28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u] + 2u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u + 1u + event[28u + event[27] + 3u + 1u + event[28u + event[27] + 3u] + 1u] + 2u]));
 }
 /**
  * @brief Get field vendor_specific_metadata_value_length from event GATTSERVICE_SUBEVENT_PACS_CLIENT_PACK_RECORD
